Update script.js
Added quotes array, colors array, and several functions so that random quotes and background colors will be displayed on button click. I also added a timer to change the quotes at set intervals but I am currently still working on a solution that will reset the timer when the button is pressed.